Fix NPE when attempting to perform search
The newLdaptiveSearchOperation doesn't set the LDAP ConnectionFactory.  This causes a NPE when attempting
to make use of LDAP groups to determine if the end-user is authorized to use the cas-management application.
The NPE is thrown when the SearchOperation attempts to execute the search, and is unable to create an LDAPConnection.